Add change column order operation

+++ src/main/java/com/likeTrello/board/model/tasks/model/Task.java
@@ -18,7 +18,7 @@ public class Task {
     private Long id;
 
     @Column(name = "task_order")
-    private Long taskOrder;
+    private Integer taskOrder;
 
     @Column(name = "task_name")
     private String taskName;
